storage problem fixed and download post as pdf

+++ controllers/post.js
@@ -2,6 +2,14 @@ const Post = require("../models/post");
 const { post } = require("../routes/post");
 const { validationResult } = require("express-validator");
 const { formatISO9075 } = require("date-fns");
+const pdf = require("pdf-creator-node");
+
+const fs = require("fs");
+const expath = require("path");
+
+const fileDelete = require("../utils/fileDelete");
+const path = require("path");
+const { type } = require("os");
 
 exports.createPost = (req, res, next) => {
   const { title, description } = req.body;
@@ -34,15 +42,6 @@ exports.createPost = (req, res, next) => {
       const error = new Error("Something wrong bby!");
       return next(error);
     });
-
-  //   console.log(`Title value is ${title} & description is ${description}`);
-  // posts.push({
-  //     id : Math.random(),
-  //     title,
-  //     description,
-  //     photo
-  // });
-  // res.redirect("/");
 };
 
 exports.renderCreatePage = (req, res) => {
@@ -157,6 +156,7 @@ exports.updatePost = (req, res, next) => {
       post.title = title;
       post.description = description;
       if (image) {
+        fileDelete(post.imgUrl);
         post.imgUrl = image.path;
       }
       return post.save().then((result) => {
@@ -174,7 +174,14 @@ exports.updatePost = (req, res, next) => {
 
 exports.deletePost = (req, res, next) => {
   const { postId } = req.params;
-  Post.deleteOne({ _id: postId, userId: req.user._id })
+  Post.findById(postId)
+    .then((post) => {
+      if (!post) {
+        return res.redirect("/");
+      }
+      fileDelete(post.imgUrl);
+      return Post.deleteOne({ _id: postId, userId: req.user._id });
+    })
     .then(() => {
       console.log("post deleted");
       res.redirect("/");
@@ -182,6 +189,71 @@ exports.deletePost = (req, res, next) => {
     .catch((err) => {
       console.log(err);
       const error = new Error("Something wrong bby!");
+      return next(error);
+    });
+};
+
+exports.savePostAsPdf = (req, res, next) => {
+  const { id } = req.params;
+  const templateUrl = `${expath.join(
+    __dirname,
+    "../views/template/template.html"
+  )}`;
+  const html = fs.readFileSync(templateUrl, "utf8");
+
+  const options = {
+    format: "A3",
+    orientation: "portrait",
+    border: "10mm",
+    header: {
+      height: "45mm",
+      contents:
+        '<h3 style="text-align: center;">PDF Download From Blog.io</h3>',
+    },
+    footer: {
+      height: "28mm",
+      contents: {
+        first: "Cover page",
+        contents:
+          '<span style="color: #444; text-align: center;">@johathan.mm</span>',
+      },
+    },
+  };
+
+  Post.findById(id)
+    .populate("userId", "email")
+    .lean()
+    .then((post) => {
+      const date = new Date();
+      const pdfSaveUrl = `${expath.join(
+        __dirname,
+        "../public/pdf",
+        date.getTime() + ".pdf"
+      )}`;
+      const document = {
+        html,
+        data: {
+          post,
+        },
+        path: pdfSaveUrl,
+        type: "",
+      };
+      pdf
+        .create(document, options)
+        .then((result) => {
+          console.log(result);
+          res.download(pdfSaveUrl, (err) => {
+            if (err) throw err;
+            fileDelete(pdfSaveUrl);
+          });
+        })
+        .catch((error) => {
+          console.error(error);
+        });
+    })
+    .catch((err) => {
+      console.log(err);
+      const error = new Error("Something Wrong Bby");
       return next(error);
     });
 };
